refactor storyblok richtext
